wip: api specs integration
