+++ content/src/controller/error-handler.ts
@@ -2,7 +2,7 @@ import { IHttpServerComponent } from '@well-known-components/interfaces'
 import { InvalidRequestError } from '../types'
 
 export async function errorHandler(
-  ctx: IHttpServerComponent.DefaultContext<object>,
+  _ctx: IHttpServerComponent.DefaultContext<object>,
   next: () => Promise<IHttpServerComponent.IResponse>
 ): Promise<IHttpServerComponent.IResponse> {
   try {
@@ -12,13 +12,11 @@ export async function errorHandler(
       return Promise.resolve({
         status: 400,
         body: {
-          error: 'Bad request',
-          message: error.message
+          error: error.message
         }
       })
     }
 
-    console.log(`Error handling ${ctx.url.toString()}: ${error.message}`, error)
     throw error
   }
 }
